+++ tests/Transformers/MinMaxNormalizerTest.php
@@ -2,8 +2,10 @@
 
 namespace Rubix\ML\Tests\Transformers;
 
+use Rubix\ML\Persistable;
 use Rubix\ML\Transformers\Elastic;
 use Rubix\ML\Transformers\Stateful;
+use Rubix\ML\Transformers\Reversible;
 use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Transformers\MinMaxNormalizer;
@@ -45,12 +47,14 @@ class MinMaxNormalizerTest extends TestCase
         $this->assertInstanceOf(Transformer::class, $this->transformer);
         $this->assertInstanceOf(Stateful::class, $this->transformer);
         $this->assertInstanceOf(Elastic::class, $this->transformer);
+        $this->assertInstanceOf(Reversible::class, $this->transformer);
+        $this->assertInstanceOf(Persistable::class, $this->transformer);
     }
 
     /**
      * @test
      */
-    public function fitUpdateTransform() : void
+    public function fitUpdateTransformReverse() : void
     {
         $this->transformer->fit($this->generator->generate(30));
 
@@ -68,15 +72,23 @@ class MinMaxNormalizerTest extends TestCase
         $this->assertIsArray($maximums);
         $this->assertCount(3, $maximums);
 
-        $sample = $this->generator->generate(1)
-            ->apply($this->transformer)
-            ->sample(0);
+        $dataset = $this->generator->generate(1);
+
+        $original = $dataset->sample(0);
+
+        $dataset->apply($this->transformer);
+
+        $sample = $dataset->sample(0);
 
         $this->assertCount(3, $sample);
 
         $this->assertEqualsWithDelta(0.5, $sample[0], 1);
         $this->assertEqualsWithDelta(0.5, $sample[1], 1);
         $this->assertEqualsWithDelta(0.5, $sample[2], 1);
+
+        $dataset->reverseApply($this->transformer);
+
+        $this->assertEquals($original, $dataset->sample(0));
     }
 
     /**